+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Added `ifEmpty` to print something specific for empty `Iterables`.
   * Added `map` and `cast` printers for more flexible printing.
   * Added fully configurable object / `toString()` printer.
+  * Added `indent` and `dedent` printer.
 
 ## 3.1.0